+++ visual_graph_datasets/experiments/generate_molecule_dataset_from_csv__aggregators_lee.py
@@ -1,0 +1,156 @@
+"""
+This experiment processes the aggregators dataset. This is orignally a dataset consisting of about 300k molecules 
+that are annotated with a binary classification label that identifies them as either an aggregator or a 
+non-aggregator.
+
+CHANGELOG
+
+0.1.0 - 23.02.2023 - initial version
+"""
+import os
+import pathlib
+import typing as t
+
+# from pycomex.experiment import SubExperiment
+# from pycomex.util import Skippable
+from pycomex.functional.experiment import Experiment
+from pycomex.utils import folder_path, file_namespace
+
+
+PATH = pathlib.Path(__file__).parent.absolute()
+ASSETS_PATH = os.path.join(PATH, 'assets')
+
+# == SOURCE PARAMETERS ==
+# These parameters determine how to handle the source CSV file of the dataset. There exists the possibility
+# to define a file from the local system or to download a file from the VGD remote file share location.
+# In this section one also has to determine, for example, the type of the source dataset (regression, 
+# classification) and provide the names of the relevant columns in the CSV file.
+
+# :param FILE_SHARE_PROVIDER:
+#       The vgd file share provider from which to download the CSV file to be used as the source for the VGD
+#       conversion. 
+FILE_SHARE_PROVIDER: str = 'main'
+# :param CSV_FILE_NAME:
+#       The name of the CSV file to be used as the source for the dataset conversion.
+#       This may be one of the following two things:
+#       1. A valid absolute file path on the local system pointing to a CSV file to be used as the source for
+#       the VGD conversion
+#       2. A valid relative path to a CSV file stashed on the given vgd file share provider which will be
+#       downloaded first and then processed.
+CSV_FILE_NAME: str = os.path.join(ASSETS_PATH, 'aggregators_lee.csv')
+# :param INDEX_COLUMN_NAME:
+#       (Optional) this may define the string name of the CSV column which contains the integer index
+#       associated with each dataset element. If this is not given, then integer indices will be randomly
+#       generated for each element in the final VGD
+INDEX_COLUMN_NAME: t.Optional[str] = None
+# :param SMILES_COLUMN_NAME:
+#       This has to be the string name of the CSV column which contains the SMILES string representation of
+#       the molecule.
+SMILES_COLUMN_NAME: str = 'smiles'
+# :param TARGET_TYPE:
+#       This has to be the string name of the type of dataset that the source file represents. The valid 
+#       options here are "regression" and "classification"
+TARGET_TYPE: str = 'classification'
+# :param TARGET_COLUMN_NAMES:
+#       This has to be a list of string column names within the source CSV file, where each name defines 
+#       one column that contains a target value for each row. In the regression case, this may be multiple 
+#       different regression targets for each element and in the classification case there has to be one 
+#       column per class.
+TARGET_COLUMN_NAMES: t.List[str] = ['nonaggregator', 'aggregator', ]
+
+# == DATASET PARAMETERS ==
+# These parameters control aspects of the visual graph dataset creation process. This for example includes 
+# the dimensions of the graph visualization images to be created or the name of the visual graph dataset 
+# that should be given to the dataset folder.
+
+# :param DATASET_CHUNK_SIZE:
+#       This number will determine the chunking of the dataset. Dataset chunking will split the dataset
+#       elements into multiple sub folders within the main VGD folder. Especially for larger datasets
+#       this should increase the efficiency of subsequent IO operations.
+#       If this is None then no chunking will be applied at all and everything will be placed into the
+#       top level folder.
+DATASET_CHUNK_SIZE: t.Optional[int] = None
+# :param DATASET_NAME:
+#       The name given to the visual graph dataset folder which will be created.
+DATASET_NAME: str = 'aggregators_binary'
+# :param IMAGE_WIDTH:
+#       The width molecule visualization PNG image
+IMAGE_WIDTH: int = 1000
+# :param IMAGE_HEIGHT:
+#       The height of the molecule visualization PNG image
+IMAGE_HEIGHT: int = 1000
+# :parm DATASET_META:
+#       This dict will be converted into the .meta.yml file which will be added to the final visual graph dataset
+#       folder. This is an optional file, which can add additional meta information about the entire dataset
+#       itself. Such as documentation in the form of a description of the dataset etc.
+DATASET_META: t.Optional[dict] = {
+    'version': '0.1.0',
+    # A list of strings where each element is a description about the changes introduced in a newer
+    # version of the dataset.
+    'changelog': [
+        '0.1.0 - 29.01.2023 - initial version'
+    ],
+    # A general description about the dataset, which gives a general overview about where the data was
+    # sampled from, what the input features look like, what the prediction target is etc...
+    'description': (
+        'large dataset consisting of organic compounds which are divided into two classes: aggregators '
+        'and non-aggregators.'
+    ),
+    # A list of informative strings (best case containing URLS) which are used as references for the
+    # dataset. This could for example be a reference to a paper where the dataset was first introduced
+    # or a link to site where the raw data can be downloaded etc.
+    'references': [
+        'Library used for the processing and visualization of molecules. https://www.rdkit.org/',
+    ],
+    # A small description about how to interpret the visualizations which were created by this dataset.
+    'visualization_description': (
+        'Molecular graphs generated by RDKit based on the SMILES representation of the molecule.'
+    ),
+    # A dictionary, where the keys should be the integer indices of the target value vector for the dataset
+    # and the values should be string descriptions of what the corresponding target value is about.
+    'target_descriptions': {
+        0: 'one-hot: non-aggregator class',
+        1: 'one-hot: aggregator class'
+    }
+}
+GRAPH_METADATA_CALLBACKS = {
+    'smiles': lambda mol, data: data['smiles'],
+}
+
+# == EVALUATION PARAMETERS ==
+# These parameters control the evaluation process which included the plotting of the dataset statistics 
+# after the dataset has been completed for example.
+
+# :param EVAL_LOG_STEP:
+#       The number of iterations after which to print a log message
+EVAL_LOG_STEP = 100
+# :param NUM_BINS:
+#       The number of bins to use for the histogram 
+NUM_BINS = 10
+# :param PLOT_COLOR:
+#       The color to be used for the plots
+PLOT_COLOR = 'gray'
+
+# == EXPERIMENT PARAMETERS ==
+
+experiment = Experiment.extend(
+    'generate_molecule_dataset_from_csv.py',
+    base_path=folder_path(__file__),
+    namespace=file_namespace(__file__),
+    glob=globals(),
+)
+
+
+# ~ Adding filters to the dataset processing step
+# By adding these specific filters to the pre-processing of the dataset we implement the same processing
+# steps described in the original paper which introduces this dataset.
+
+def max_graph_size(mol, data):
+    """
+    We want to filter the very big molecules because we will likely not be able to properly visualize
+    those for the explanations anyways.
+    """
+    return len(mol.GetAtoms()) >= 100
+
+
+experiment.run_if_main()